Adjusts

+++ backend/payment/src/main/java/dev/alexguesser/payment/application/usecase/input/ProcessPaymentInput.java
@@ -1,10 +1,9 @@
 package dev.alexguesser.payment.application.usecase.input;
 
-import java.math.BigDecimal;
 import java.util.UUID;
 
 public record ProcessPaymentInput(
         UUID rideId,
-        BigDecimal amount
+        double amount
 ) {
 }
